% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{hyperband_n_configs}
\alias{hyperband_n_configs}
\title{Hyperband Configs}
\usage{
hyperband_n_configs(r_min, r_max, eta)
}
\arguments{
\item{r_min}{(\code{numeric(1)})\cr
Lower bound of budget parameter.}

\item{r_max}{(\code{numeric(1)})\cr
Upper bound of budget parameter.}

\item{eta}{(\code{numeric(1)})\cr
Fraction parameter of the successive halving algorithm: With every stage the
configuration budget is increased by a factor of \code{eta} and only the best
\code{1/eta} points are used for the next stage. Non-integer values are supported,
but \code{eta} is not allowed to be less or equal 1.}
}
\value{
\code{integer(1)}
}
\description{
Calculates how many different configurations are sampled.
}
