% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3hyperband-package}
\alias{mlr3hyperband}
\alias{mlr3hyperband-package}
\title{mlr3hyperband: Hyperband for 'mlr3'}
\description{
Implements hyperband method for hyperparameter tuning. 
    Various termination criteria can be set and combined. The class 'AutoTuner'
    provides a convenient way to perform nested resampling in combination with
    'mlr3'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mlr3tuning.mlr-org.com}
  \item \url{https://github.com/mlr-org/mlr3hyperband}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3hyperband/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Gruber \email{gruber_sebastian@t-online.de} (0000-0001-9754-0393)

Authors:
\itemize{
  \item Bernd Bischl \email{bernd_bischl@gmx.net} (0000-0001-6002-6980)
}

}
