% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nds_selection.R
\name{nds_selection}
\alias{nds_selection}
\title{Best points w.r.t. non dominated sorting with hypervolume contrib.}
\value{
\code{integer()}
}
\description{
Select best subset of points by non dominated sorting with
hypervolume contribution for tie breaking. Works on an arbitrary dimension
of size two or higher. Returns a vector of indices of selected points.
}
\section{Parameters}{

\describe{
\item{\code{points}}{\code{matrix()}\cr
Numeric matrix with each column corresponding to a point.}
\item{\code{n_select}}{\code{integer(1)}\cr
Amount of points to select.}
\item{\code{ref_point}}{\code{integer()}\cr
Reference point for hypervolume.}
\item{\code{minimize}}{\code{logical()}\cr
Should the ranking be based on minimization? (Single bool for all dimensions,
or vector of bools with each entry corresponding to each dimension).}
}
}

