% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_TuningInstance.R
\name{autoplot.TuningInstance}
\alias{autoplot.TuningInstance}
\title{Plot for TuningInstance}
\usage{
\method{autoplot}{TuningInstance}(
  object,
  type = "boxplot",
  measure = NULL,
  hbbracket = NULL,
  params = NULL,
  facet_per_bracket = TRUE,
  ...
)
}
\arguments{
\item{object}{(\link[mlr3:TuningInstance]{mlr3::TuningInstance}).}

\item{...}{(\code{any}):
Additional arguments, passed down to the respective \code{geom}.}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Generates plots for \link[mlr3tuning:TuningInstance]{mlr3tuning::TuningInstance}, depending on argument \code{type}:
\itemize{
\item \code{"boxplot"} (default): Boxplots of evaluated performance for the
respective hyperparameter configurations if tuning was performed by \link[mlr3hyperband:TunerHyperband]{mlr3hyperband::TunerHyperband}. Boxplots are shown in facets per bracket. Additionaly, the budget spent is shown.
}
}
\examples{
library(mlr3hyperband)

task = tsk("pima")
learner = lrn("classif.xgboost")
resampling = rsmp("holdout")
measure = msr("classif.acc")

ps = ParamSet$new(
   params = list(
      ParamInt$new("nrounds", lower = 1L, upper = 270L, tags = "budget"),
      ParamInt$new("max_depth", lower = 1, upper = 100)
   )
)
term = term("evals", n_evals = 100000L)
inst = TuningInstance$new(task, learner, resampling, measure, ps, term)
tuner = tnr("hyperband", eta = 3L)

head(fortify(object))
autoplot(object)
}
