% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3hyperband-package}
\alias{mlr3hyperband}
\alias{mlr3hyperband-package}
\title{mlr3hyperband: Hyperband for 'mlr3'}
\description{
Implements hyperband method for hyperparameter
    tuning.  Various termination criteria can be set and combined. The
    class 'AutoTuner' provides a convenient way to perform nested
    resampling in combination with 'mlr3'. The hyperband algorithm was
    proposed by Lisha Li, Kevin Jamieson, Giulia DeSalvo, Afshin
    Rostamizadeh and Ameet Talwalkar (2018) <arXiv:1603.06560>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mlr3hyperband.mlr-org.com}
  \item \url{https://github.com/mlr-org/mlr3hyperband}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3hyperband/issues}
}

}
\author{
\strong{Maintainer}: Marc Becker \email{marcbecker@posteo.de} (\href{https://orcid.org/0000-0002-8115-0400}{ORCID})

Authors:
\itemize{
  \item Sebastian Gruber \email{gruber_sebastian@t-online.de} (\href{https://orcid.org/0000-0002-8544-3470}{ORCID})
  \item Jakob Richter \email{jakob1richter@gmail.com} (\href{https://orcid.org/0000-0003-4481-5554}{ORCID})
  \item Julia Moosbauer \email{ju.moosbauer@googlemail.com} (\href{https://orcid.org/0000-0002-0000-9297}{ORCID})
  \item Bernd Bischl \email{bernd_bischl@gmx.net} (\href{https://orcid.org/0000-0001-6002-6980}{ORCID})
}

}
