% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nds_selection.R
\name{nds_selection}
\alias{nds_selection}
\title{Select best subset of points by non dominated sorting with
hypervolume contribution for tie breaking}
\usage{
nds_selection(points, n_select, ref_point = NULL, minimize = TRUE)
}
\arguments{
\item{points:}{matrix with each column corresponding to a point}

\item{n_select:}{Amount of points to select (integer(1L))}

\item{ref_point:}{Reference point for hypervolume (integer())}

\item{minimize:}{Should the ranking be based on minimization (single bool
for all dimensions, or vector of bools corresponding to the dimensions)}
}
\value{
indeces of selected points
}
\description{
Select best subset of points by non dominated sorting with
hypervolume contribution for tie breaking
}
