% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TunerHyperband.R
\docType{data}
\name{TunerHyperband}
\alias{TunerHyperband}
\alias{mlr_tuners_hyperband}
\title{TunerHyperband}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Tuner}.}
\description{
Subclass for hyperband tuning.

The grid is constructed as a Cartesian product over discretized values per parameter,
see \link[paradox:generate_design_grid]{paradox::generate_design_grid}.
The points of the grid are evaluated in a random order.

In order to support general termination criteria and parallelization,
we evaluate points in a batch-fashion of size \code{batch_size}.
Larger batches mean we can parallelize more, smaller batches imply a more fine-grained checking
of termination criteria.
}
\section{Construction}{
\preformatted{tuner = TunerHyperband$new(eta = 3L)
}
\itemize{
\item \code{eta} :: \code{integer(1)} \cr
Fraction parameter of the successive halving algorithm: With every step
the configuration budget is increased by a factor of \code{eta} and only the
best \code{1/eta} configurations are used for the next step.
}
}

\examples{
# see ?Tuner
}
\references{
\url{https://arxiv.org/abs/1603.06560}
}
\concept{Tuner}
\keyword{datasets}
